Auto generated todo list for rubocop
Created by executing `rubocop --auto-gen-config`

Currently since the .rubocop.yml file is empty it uses the default.yml settings as the default for the style-guide. Custom rules can be added into the local .rubocop.yml and it will override the default.

`inherit_from: .rubocop_todo.yml` currently ignores the violations reported in .rubocop_todo.yml.

Next step is to remove the exclusions in the todo file and fix the violations. This way all the corrections regarding a particular style can be added in a single commit. Repeat this till all exclusions are fixed.